requirements.txt
backcall==0.2.0
click==8.0.3
colorama==0.4.4
debugpy==1.5.1
decorator==5.1.1
entrypoints==0.3
Flask==2.0.2
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
ipykernel==6.6.1
ipython==7.31.0
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.3
jupyter-client==7.1.0
jupyter-core==4.9.1
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
nest-asyncio==1.5.4
parso==0.8.3
pickleshare==0.7.5
prompt-toolkit==3.0.24
psycopg2==2.9.3
Pygments==2.11.2
python-dateutil==2.8.2
python-dotenv==0.19.2
pywin32==303
pyzmq==22.3.0
six==1.16.0
SQLAlchemy==1.4.29
tornado==6.1
traitlets==5.1.1
wcwidth==0.2.5
Werkzeug==2.0.2
